gdbstub.h: Fix GDB Interrupt Packet handling
Whenever an interrupt signal is fired from within GDB (Ctrl+C for example), it sends a simple packet w/the value 0x03; to handle such case we expand the current impl. w/a user designated option (Could be the user-defined stop method) and we send a SIGINT signal ("S02", but could be any other permitted value).

I'm not sure how the maintainers want to handle the user-defined option to handle gdb interrupts (I use gdb->config.stop(gdb->config.user_data) because it works on my current setup) so if it needs something more generic we can change it.